pan_os_upgrade/requirements.txt
ansible==9.2.0
ansible-compat==4.1.11
ansible-core==2.16.3
ansible-lint==6.22.2
asttokens==2.4.1
attrs==23.2.0
bcrypt==4.0.1
black==24.1.1
boto3==1.26.78
botocore==1.29.164
bracex==2.4
certifi==2024.2.2
cffi==1.15.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
cryptography==41.0.2
executing==2.0.1
filelock==3.13.1
icecream==2.1.3
idna==3.6
Jinja2==3.1.3
jmespath==1.0.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
mypy-extensions==1.0.0
orionsdk==0.4.0
packaging==23.2
pan-os-python==1.8.1
pan-python==0.17.0
paramiko==2.12.0
pathspec==0.12.1
platformdirs==4.2.0
pycparser==2.21
Pygments==2.17.2
PyNaCl==1.5.0
python-dateutil==2.8.2
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.33.0
requests==2.31.0
resolvelib==1.0.1
rich==13.7.0
rpds-py==0.17.1
ruamel.yaml==0.18.5
ruamel.yaml.clib==0.2.8
s3transfer==0.6.1
six==1.16.0
subprocess-tee==0.4.1
urllib3==1.26.18
wcmatch==8.5
xmltodict==0.13.0
yamllint==1.34.0
